server/requirements.txt
aerospike==15.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
boto3==1.34.113
botocore==1.34.113
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
google==3.0.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
load-dotenv==0.1.0
MarkupSafe==3.0.2
mergedeep==1.3.4
multidict==6.1.0
openai==0.27.8
propcache==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.1.3
yarl==1.18.3
